Make dd a little more useful for natives
If you specify dd $a, and $a is a native int or str, then you only get
to see the value *without* the name.  This is annoying while debugging.
To make this easier, you can now specify natives as a named parameter.

  my int $a = 42; dd :$a;   # Int a = 42

and by the magic of the :$var notation, we can now see the name.