Add IO::Spec::Any
In the old design an IO::Spec::Win32 isa IO::Spec::Unix, this violation of the Liskov substitution principple causes various issues. This patch adds a IO::Spec::Any role containing the shared interface, as well as shared implementations. I would have preferred to add this to IO::Spec itself, but that is not possible while it's exposing a (documented) class method (select).