Make xx quite a bit faster
- 42 xx 100      # no change, was already optimized
- rand xx 100    # about 3.5x as fast
- 42 xx *        # about 5x as fast
- rand xx *      # about 1.5x as fast

The fixed lenghts xx completely build the list in memory.  For large
values, that may not be such a good idea.  On the other hand, you would
probably use a * in those cases anyway.  Should this turn out to be a
problem, then maybe we can switch to an iterator type approach if N > X.
Of course, in the "push-all" case, there would be only a performance
degredation, so I'm not so sure about such an aproach anyway.  The
* cases are implemented now as proper Iterators.