Prevent creation of Capture from lazy list.
Rather than just hanging (which was a slight improvement from the
pre-GLR behavior, which was to give back an empty Capture, meaning
you could get silent bogus matches on an empty array unpack in a
signature). Thus this addresses RT #123581. In a future Perl 6, we
may find a way to allow Capture to contain infinite things. But not
for 6.christmas.