Introduce .head($n)/.tail($n)
Inspired by List::Slice, mentioned in

  http://blogs.perl.org/users/preaction/2015/10/listslice---slice-like-operations-for-lists.html

I figured we could use a .head and .tail as well in Perl 6.  Basically,
the foo.head($n) is just another way of saying foo[^$n]: it should just
reify the elements that are needed.  Although semantically foo.tail($n)
is the same as foo[*-10..*], it *will* also reify all elements of the
list, *but* it will only keep $n elements around to be passed on when
the source list is exhausted.  This potentially uses a *lot* less memory.