Let NC map int to longlong
While `int` in nativecall declarations might be somewhat confusing,
mapping it to the same thing as p6-land native `int` would be
consistent - in normal situation, `int` would just be an alias for
`int64` and `num` for `num64`.

On 'mainstream' architectures, this patch should be enough to do it,
possibly there will be a better solution in the future.

Closes #572.