builder: add capability to work with staging directories
To allow building for a staging root which might differ from host and tools to be used,
be more careful about the PATH's to be used. This will allow cross-compiling rakudo
using two builds of nqp: a native one in SDKROOT and a cross-compiled one in SYSROOT.
Using host one remains untouched when run without --sysroot or --sdkroot, respectively.

Signed-off-by: Jens Rehsack <sno@netbsd.org>