Make List.sort 30% faster
And also not lazy: the thing is that there is little point in laziness
on the result of a sort: all data is in memory already, so making things
lazy just adds overhead.  The only case this is slower, is for the
sink-all and count-only iterator cases.  But if you doing a sort in
sink context, or counting number of elements in a sort, you're doing it
wrong anyway.