Don't try to load precomp files after changing the precomp store
In the following sequence:
use Test;
use lib 't/spec';
use Test::Utils;
after the 'use lib' we not only may no longer precompile modules, but also may
not load precompiled modules. That's because the precompiled Test::Utils may
link to a different precompiled version of Test. On loading Test::Utils this
Test module would be loaded and conflict with the already loaded Test.

The moral of the story is: to take most advantage of precompilation, try to set
up your repository chain before starting to load modules.