Fix initial value for index variable
This index variable was wrongly initialized to 0 with commit 019a7ff7.

Said commit reverted commit f0c6a02c, but in this specific case some
related code was changed in the meantime with commit 3bbc9222. That
lead to a situation where the corresponding change to the 'while'
clause 4 lines below was not reverted -- only the initialization of $i.

This lead to some fallout on rakudo.jvm: Even '1.sort' resulted in
a Null Pointer Exception. On rakudo.moar there seemed to be no
spectest fallout.