Use the precomp repository of the head of the chain on installation
The need method is passed the precomp repository by its predecessor in the
chain. The install method is called directly but it still needs to use the
head's precomp repository as the full chain is used to resolve
dependencies during precompilation and we want to install into the store
that is actually used for loading.