Precompile on-demand in Repository::Installation, too.
CompUnit::Repository::Installation needs to precompile on-demand (when loading
a module) in addition to on installation. Consider:
use lib "lib"; # prepends a Repository::FileSystem to the repo chain
use Test; # installed in site, therefore not in lib's precomp store
          # therefore loaded from source
use Test::Mock; # loaded from lib/ therefore precompiled on-demand.

Since Test::Mock uses Test, the latter will be precompiled, too when
precompiling Test::Mock. But Test was already loaded from source before
precompilation of Test::Mock so when we finally load the latter, we get
a conflict.

If you precompile on-demand, you have to do it everywhere.