Fix various data loss bugs with async I/O.
In some cases, async sockets and processes could lose data from the
end of the stream. Further, there was a data race in the sequencing
code, which was duplicated in a couple of places also. This factors
the sequence logic out to a single place, removes the data race, and
uses protocol changes in Moar/JVM backend code to fix the data loss
issue.