Check precomp files of dependencies against their source
Before loading a precomp file, we already checked if its newer than all its
dependencies. But we only checked the precomp files of the dependencies.

Now we also check if those precomp files are actually newer than their sources.
This fixes strange errors when chaning the source of a dependency of a
precompiled module.