Make open(:bin) do something useful
Previously, :bin was simply ignored as even without setting the
encoding in Rakudo, MoarVM would default to utf8.

By spec, :bin is supposed to allow binary line-by-line access via
.get and .lines.

Using the utf8-c8 encoding and mixing in a version of .get that calls
.encode on its return value makes this happen.

The whole thing is something of a hack, but for now it's probably
Good Enough.