Use distribution's name to find binary resources
Currently, when a distribution has binary/script files, we install them
as resources owned by the distribution, as well as wrapper scripts that
the shell can find via PATH that use the CompUnit API to dispatch to
the corresponding resource.

Some distributions install multiple binaries; for example, panda
installs panda, panda-build, panda-test, etc.  The wrapper scripts
currently use the basename of the wrapper script to locate the
distribution that contains the corresponding binary resource, rather
than the name of the distribution that binary is a part of.

This change makes wrapper scripts use the distribution name to find dist
metadata, which will then be used to locate the resource for the actual
binary.