Fix spurious file locking problems with precomp store
File locking is a mess. Because it's racy, we cannot rely on upgrading a
shared lock to an exclusive lock, so we have to take an exclusive lock
even for reading precomp files.

We also now only lock a file once and count how often we tried locking
the same file, so we only unlock after the last locker set it free.

Lastly we re-check the existence of a precomp file after loading failed
and we locked the store again for precompilation, because in between,
someone else may have precompiled and we could only make matters worse
if we precompile again.