Expand the non-punning methods list for roles.
These methods will now not try to pun the role, but work on the type
object of the role itself. This will fix things like === doing .WHICH
and so trying to pun the role, and similar for .gist, .perl and some
others. This was mostly an issue because a role with stubbed methods
cannot be punned (and it would be rather odd to make it do so).