Fix use lib accidentally cutting the $*REPO chain short
If one used "use lib" to add a repository that has already been in the $*REPO
chain, it would get moved to the front and the part of the chain pointed to by
its .next-repo was lost.

We now properly remove it from the chain before putting it to the front.

Thanks to bartonlin++ for reporting!