Fix $*REPO chain broken by use lib
If a "use lib" statement occurs before $*REPO is initialized and the
repository would already be in the $*REPO chain, it would never get its
.next-repo attribute set. This is caused by repository-for-spec caching
the created repositories, so only the first access determined the
next-repo. The "use lib" implementation doesn't pass a :next-repo, so it
was always undefined. OTOH if we just blindly overwrite .next-repo, we
may create loops in the chain.

Fixed doing the full "cut out of the repo-chain and re-insert properly"
dance that use-repository does, even when setting up the initial chain.

Fixes lots of "Cannot find Test in..." errors.
Thanks to Tux++ for the report!