Make Pair a value type when it can
In most cases, Pair's are value types.  The key of a Pair is constant.
The value usually is as well.  You need to do special stuff like:

  a => my $ = 42

to make it changeable.  Only in that case, will Pair.WHICH now revert
to the old .WHICH behaviour.  In any other case, it will return a
value typed .WHICH, making all Pairs with a given key and value give
the same .WHICH value.  Making them useable in Sets, Bags and Mixes,
and have them usable as a key in a typed hash.