Fix passing the repo-chain to the precomp process
We have to make sure that the process precompiling one of our dependencies
uses exactly the same repository chain as the process trying to load the module.
For that we skip the normal repository setup and instead serialize the parent's
repositories into an environment variable and restore them later.

The special purpose repository implementations AbsolutePath, NQP and Perl5 did
not register their path-spec prefixes properly with the RepositoryRegistry, so
on deserialization we believed them to be FileSystem repositories, thereby
altering the repository chain and ending up generating different precomp-ids.