Respect multi's sigs in role composition.
This means that a multi foo(Int $x) { ... } in a role will require an
exactly matching candidate be provided. Prior to this, you'd get an
ambiguous dispatch. Further, a multi in a role with a matching name
and signature with one in a class will now hide the one from the role;
this was also previously an ambiguous dispatch.