fix for 125769, binding in sink context throws
The wanter now notices such binds at the ends of blocks and
install an extra True value afterwards to shield the binding
from autothrow semantics.  In addition, if/else/with/without
now propagage sinkness properly to the then/else clauses, which
exposed the fact that lots of instances of $_ weren't wanted yet.