discourage casual use of EVAL
It is now by default illegal in a normal program to use the EVAL function.
This may be overridden in any of several ways:
    use MONKEY-SEE-NO-EVAL;
    use MONKEY;  # shortcut that turns on all MONKEY pragmas
    use Test;
    setenv RAKUDO_MONKEY_BUSINESS=1

The MONKEY pragmas always take precedence, so regardless of use of Test or
environment variables, a lexical scope may turn the capability on and off.
So something like an IRC bot or REPL may default to allowing EVAL, but it's
still controllable either way within the actual code.

Several other MONKEY names have been reserved for similar safety, in addition
to the original MONKEY-TYPING.  'use MONKEY-GUTS' is now a synonym for 'use nqp.'
We might restrict things like .^set_value with 'use MONKEY-BRAINS'.

Regular expressions record their monkey state at compile time and use that
info to decide whether to allow interpolations later at run time.