Load dependencies of a precomp file directly instead of searching
Since for every precompiled file we already record the ids of all recursive
dependencies, we can use that information for loading those dependencies,
instead of searching all over again.

This is also an important step to get rid of RepositoryRegistry.load_module

Many thanks to jnthn++ for pointing out this possibility!