Use empty string for undefined library names
Some C library implementations, for example, glibc, allow a NULL string
to be passed to their dynamic library loading function (dlopen in the
case of glibc).  This returns a handle for the currently executing
image, and is a pretty popular way to access libc routines dynamically.
This change makes it possible to use `is native(Str)` once again