while/until loops can now return values
A loop that has KEEP or UNDO, or that is used in a spot indicating
that a return value is expected (such as in parens), will now be
translated to a map (well, a for loop) that knows how to deal with
blocks returning values correctly.  In the case that a loop has a
KEEP/UNDO but is not in a context wanting the return value, it still
does the transformation to all KEEP/UNDO to have a value to test, but
then sinks the results of the loop as a whole.

Fixes #126005