METAOP_ASSIGN shall use .DEFINITE instead of .defined
this will correctly let Failure objects in a variable
blow up when the variable is +='d or so, rather than
defusing and replacing the Failure object.

The optimization for METAOP_ASSIGN in the optimizer has
sadly become a casualty, as it unexplicably leads to a
"Cannot modify immutable *" error in many places.

That's also why the implementation uses if/else rather
than a ?? !!. That leads to the same error.