Promise.[anyof|allof] no longer break.
They now only care about promises completing *somehow* rather than
about the nature of that completion. For `anyof` you'll be doing some
follow-up work to get a result anyway, and when you care about getting
all of a set of results, you just use `await` with a list of promises.
This makes anyof/allof a little easier to work with for their common
uses, which is waiting until one/all of a set of things have taken
place, when you'll then go on to be interested in them individually.