Fix `make install` when nqp is distro-installed
During the installation process, tools/build/install-core-dist.pl is called to
precompile the bundled non-CORE.setting libraries such as Test.pm. The way this
was initially coded, it pulled the install path from NQP's build-time
configuration, which will be a root-owned filesystem path if NQP was installed
via a system package manager.

Other parts of the CompUnit code will see these paths don't exist, and attempt
to helpfully call `mkdir` on them which brings the whole process to a messy end.
This part usually works fine when building everything from a git checkout,
because Moar/NQP/Rakudo all default to an installation prefix the user has write
access to.

This patch fixes that step of the build process by passing the makefile's path
prefixes to install-core-dist.pl (ensuring precompilation output goes to the
right place), and modifying it to *only* use that path (ensuring it doesn't try
to mkdir where it shouldn't). This isn't a 100% solution, but it fixes enough
to unblock packaging work on a few distros.

Thanks to mst++ for doing almost all of the legwork here, crux++ for an initial
patch that gave us a few pointers to work from, and nine++ for giving us some
Boxing Day fun ☺