reimplement value loops using Seq.from-loop
This fixes the infinite loop on 'loop', as well as allowing repeat loops
to return values too.  This patch also removes the FAKE_MACRO hack, and
gets us closer to how jnthn++ wanted it to work originally.  (It does
not yet fix the outer bug of #127069, but likely makes it easier to fix.)