Add IO::CatPath/CatHandle, remove IO::ArgFiles
Basically generalize and expand the functionality of IO::ArgFiles
into a generic IO::CatPath and IO::CatHandle class.  Features:

- $*ARGFILES is initialized using IO::CatHandle the same way it was
  before (assume strings are file names, use $*IN if no strings given)
- apart from .lines, can now also do .words, .comb, .split, and
  .Supply or any other IO::Handle method that relies on .read or
  .readchars.
- List.IO creates IO::CatPath, for easy streaming of multiple files.

Done as a single commit, that can be reverted/reapplied/trashed once
we figure out how we're going to do new features and we reached a
decision on this.  FWIW, I intended this to happen before 6.c, but
LPW Keynote and ~10 days of illness made this impossible  :-(