Fix for RT #127093
It turned out that somehow the default value of the ContainerDescriptor
for a typed, shaped array, wasn't set.  By setting it explicitely, the
described problem goes away.  Perhaps this has something to do with the
fact that you cannot set default values on shaped arrays yet.

Jonathan: I was wondering whether the .new of ContainerDescriptor should
not just set the default attribute with the :of value if no :default is
specified.  OTOH, I'm not sure how to detect :default not being passed
in nqp.