Make error messages on "is default/dynamic" better
This is really the result of trying to make "is default/dynamic" work
on shaped arrays.  This has failed, because at the moment the trait
code is run, the shaped array does not have a descriptor initialized
yet.  Which means we're trying to look up attributes in a type object,
and thus fail.