Fix file ids changing on reinstallation of a module
Precompiled modules are checked against their source files on loading to detect
outdated precomp files. This is done by recording the source file's id. Thus
this id must be stable and survive reinstallation of a module. Otherwise the
precompilation file's modification time will be compared to an outdated source
file.

Note that this also means that the source file's path must be stable. A patch
to panda will fix those paths.