Make RAKUDO_MODULE_DEBUG show times in msecs
And only the difference between this and the previous step.  Getting
Perl6/Moduleloader.nqp to participate in that program, turned out to
be something of a bridge too far: there is no setting loaded at that
point, and the sprintf handler is also not initialized yet.  So we
don't show timings for loading the setting anymore.