Fix rare "duplicate definition of symbol" errors
When an EXPORT or EXPORTHOW routine assigns a symbol as in:
EXPORTHOW::<grammar> = TracedGrammarHOW;
instead of binding it, the test in Perl6::ModuleLoader::merge_symbols for
equivalence of source and destination symbols fails because it compares
the containers of those symbols instead of the symbols themselves.

Fix by always decont'ing before comparison.

Many, many thanks to FROGGS++ for finding the final piece of this puzzle!

Fixes RT #127107