At least temporary fix for RT #127207
Seems the optimizer thought that all Num values where actual numeric
values and not abstract things like Inf or NaN.  This patch (probably
very badly) checks for that situation, and basically just emits the
compile time error if it *is* an actual literal numeric value.  Else
it just ignores, and it becomes a runtime error.

Leaving it as an exercise to the more knowledgable Optimizer.nqp devs
to make it go to the exception *and* handle the error in such a way that
it won't suggest trying to coerce the value to Int/Rat/Complex first.