Re-imagine lazy init of system dynamic vars
The old system was using MMD to select the right candidate for lazy
initialization of system dynamic vars such as $*CWD.  This however
appeared to be incompatible with the multiple settings idea of being
able to support 6.c and 6.d by switching settings.

The new system uses an internal method for registering the initialization
of a system dynamic var.  And the internal method for looking up system
dynamic vars, now uses another initial method to do the lookup and
initialization.

So, this should now be ready for settings.d