Simplify to-json proto sig
Since each of the candidates has the same handling of the possible
named variables, it would seem to me that the specificity of the
proto sig would keep it from getting optimized away.

The sig was part of the original addition to the repo, so no other
information could be gleaned as to why this specific sig was used.