Allow --named-switches anywhere in command line.
Perl6 was GNU-like in that it has '--double-dashes' and that it stops
interpreting named parameters when it encounters '--', but unlike GNU-like
parsing, it also stopped interpreting named parameters when encountering any
positional argument.

This patch makes it a bit more GNU-like by allowing named arguments after
a positional, to prepare for allowing subcommands.