make CUR.bindir overridable
This is useful when several CURs share a single bin directory on disk.
Also when we have repositories that keep the distribution data on a non-local
storage, we need to keep the wrapper scripts in a local bindir, so that these
can be in PATH.
Rakudo Star will be the first user of this funtionality.