Override HLL::Compiler.eval for multi-line input
The way this works is pretty dodgy, but it's good enough for
my experiment.  How this works is that we catch any exceptions
from evaluating the code; if there's an exception, it's an
X::Syntax::Missing, and it occurs at the end of the input string,
that means that more input is required.  Instead of just throwing
the exception at the user, we signal HLL::Compiler that we need
more input via self.needs-more-input