Just call $<nu>.ast.compile_time_value in Perl6::Actions.bare_rat_number
The value of $<nu>.ast can either be a QAST::Want or a QAST::WVal depending on
the value of the numerator. Any value of 2**31 or greater gives us a
QAST::WVal.

Previously the code called "@($<nu>.ast)[0].compile_time_value" - this works
fine for a QAST::Want, but a QAST::WVal does not have any children, and so
this would lead to an error of "Cannot find method 'compile_time_value'".

Internally QAST::Want.compile_time_value is implemented as
"self[0].compile_time_value()" so it seems like there's no good reason for us
to do same thing in Perl6::Actions. Yay polymorphism!