Improve Int %h by 20% by removing code
Why?  Because the candidates that handle the type constrained
values Hash, were less optimised.  But the only difference really
was how they tried to get at the Hash descriptor (which is the
only thing that really separates a type constrained Hash from a
"normal" Hash).  So removing the code, will revert back to the
better optimised un-constrained candidates, and thus get better
performance.