Change fake BUILD methods to SET-SELF
If a method is called BUILD, it has a special meaning for the
default way of building objects.  These methods only shared the
name, but were never intended to be called by BUILDALL.

Having a SET-SELF in a class, usually means the class is not
subclassable as is.  By making the method private, it should
hopefully optimize even better in the future