Make enum fail with a NYI error
The enum handling tries to rebless the value to the enum.  Clearly,
one cannot rebless a type object into something else.  Now, whether
it should be possible to have a type object as the value of an enum
I'm letting open for now by making it a NYI error.

Skarsnik++ for spotting