fix precompiling a file that outputs to stdout
A file that gets precompiled outputs the hash and path of the compilation
units it depends on. The process that invoked the precompilation records
it and puts it into a file. Writing to stdout can therefore mess with this
mechanism.
This patch is just a bandaid until a more robust solution arrives.