REPL6: Add rudimentary Perl 6 REPL loading support
This is the start of moving advanced REPL functionality into Perl 6.
This is to lower the barrier to entry for improving the REPL, allowing
the REPL to take advantage of more advanced facilities provided by
Perl 6, and to allow quicker iteration on changing the REPL's code.

Until the Perl 6 REPL is "finished", we'll just "use REPL" to load it
to cut down iteration times; after it's "done", we'll stick it into
CORE.setting.

For now, we'll leave all of the advanced functionality that we've
added thus far to Perl6::Compiler there; if this idea ends up being
a success, we'll strip the more advanced features out of Perl6::Compiler
and leave a rudimentary but useful REPL in case REPL.pm can't be used.