Consider precomp files from all stores
Instead of just $*REPO.precomp-store, we now look for precomp files in all
precomp stores. This allows for us to actually use the precomp files generated
during installation of a distribution as those are no longer hidden behind
the home repo's precomp store.

As the "only use the head's precomp store" rule protected us from loading
a precomp file that uses an outdated dependency, we need a different solution
for that problem. Instead, we now record the dependency specification used
during compilation into the .deps file along with $*REPO.id which contains
the state of all repositories.

Whenever $*REPO.id differs from the one in a .deps file, we know that we
have to re-run the dependency resolution. If it arrives at the same result,
we know it's safe to load the precomp file.

In a future update, we will record the updated $*REPO.id somewhere, so we
don't have to re-run the dependency resolution every time we load the module.