REPL6: Allow us to back out of botched line editor init
The REPL class adds line editor functionality by mixing in
specific roles at runtime, depending which line editor module
is available.  The previous implementation had an issue where
if a module was mixed in to the REPL object but failed to
initialize, we would end up with an improperly configured REPL.
This change fixes that by cloning the current instance, mixing
into that and initializing the line editor, and only then replacing
the actual instance with the mixed in one.

I ended up having to create an init method on REPL to properly
initialize its attributes; adding BUILD not only has complications
with the provided compiler, since it's an NQP value, but it also
prevents Completions.BUILD from being mixed in.  I'll work on a way
to fix that.