Move permutations() to operators
It doesn't belong in native_array: it was only there because it needed
native arrays, and putting it in the file made sure of that.  However,
now that I'm focusing on native arrays, it felt better to move this
"corpus librum" to a better place, later in the CORE setting.