Make sure .join doesn't die on holes in str @a
Unassigned elements in a native str array, are null_s (as in nqp::isnull_s
giving a true value).  Trying to join them with nqp::join will skip these
elements completely, which is inconsistent with other uses of join.

Rather than creating a shadow str array with "" filled in, it seemed
like a better idea to bind the empty elements to the same empty string
and then join that.  Perhaps nqp::join should do something like this
under the hood.