REPL6: Warn user when we can't set up history file
This can happen if any component on the path to the history file (other than
the history file) exists but is not a directory, or if it's not
writable.  Let's give the user a heads-up so that they have a chance to
fix the problem before they lose all their history.