Make Object:D cmp Object:D 2x as fast
The basic version was testing for Inf/-Inf.  By adding candidates
for Real/Any and Any/Real and moving the test for Inf/-Inf in there,
we no longer need to do that in the basic version.  Also added a
very cheap test for identity, so comparing with itself should be a
lot faster as well.

Also nqpified the ORDER helper sub, just in case it would make a
difference.