Implement $*DEFAULT-READ-ELEMS for reading
Inspired by https://www.mail-archive.com/perl6-language@perl.org/msg35150.html

Takes its default from the DEFAULT_READ_ELEMS environment variable,
or assumes 65536 if that cannot be found either.

Since this number will both be used for string based reading (so the
number reflects the number of chars read, which can be more than 1 byte
per char), as well as binary reading, it felt right to call it ELEMS
rather than BYTES or CHARS.

Of course, all of this is provisional, and open for debate and
bikeshedding.  In any case, yary++ for letting is know his pet peeve.