REPL6: Report if desired line editor is not present
If a line editor module is not present for loading,
the mixin function will return two falsey values; the
first is what would normally be the new REPL object, the
second is whether a problem occurred when loading the module
unrelated to its presence in the compunit loading chain (ex.
a broken installation, a missing NativeCall library, etc).
In this case, reporting that kind of problem is handled by
the mixin function, but we want to add additional reporting if
the preferred backend is not present so as to not confuse the
user when that backend has not been installed.