Bump NQP_REVISION for various MoarVM fixes/tuning.
The most important fix in here addresses a memory leak that was most
liable to show up in programs that `EVAL` a lot of code, but could
show up easily in programs with lots of medium-age objects too. Also
a bit of GC tuning that hopefully improves memory behavior, plus a
UTF-16 fix, a crash fixed when reading corrupt bytecode, and some
further details in heap snapshots.