Fix a missing pre-comp store unlock.
This could cause the store to remain locked until program exit. This
was particularly problematic on Windows parallel spectests, causing a
hang. The test harness seems to read from the processes one at a time
on Windows, blocking the output of those it's not currently reading
from. A test that needed to pre-compile something while running would
block on acquiring the lock, which was being held by another script that
would not release it until process exit. In combination, this created a
circular wait, and thus a deadlock. However, there are probably platform
independent deadlocks that could have been caused by this bug, not to
mention inefficiency.