Fix EVAL during precompilation
EVALed code is considered an independent compilation unit that gets compiled
using a new World which contains an independent serialization context. It also
gets its own QAST compiler.

When a BEGIN time EVAL is run during precompilation of a module, the
independent SC would not be written into any file but would still be referenced
from the outer compilation unit. We therefor use the outer world's
CompilationContext which contains the SC. We also need to share a couple of
other attributes that keep track of the compilation's state.

The QAST compiler's frames are shared by adding them to the hash that's used
for initializing the QAST compiler's frames lookup hash.

Fixes RT #124324