Preliminary fix for RT #127968
I'm not sure why a Channel would need to create a Supply first to be
able to create a .list.  So I re-imagined Channel.list to be a Seq
with a simple Iterator that is polling the Channel.  This causes one
Channel spectest to fail, but I feel the test is wrong as it apparently
assumes there is a Supply underlying Channel.list.  Comments / Patches
welcome!