Harden Mu.Str against moving GC.
We have various cases of this in tests:

    is $obj-a, $obj-b, 'blah';

This uses `.Str` on the two when both are instances. The default `Str`
method used .WHERE. Unfortunately, since objects move over time, the
test could fail due to an unluckily timed GC run. Now nqp::objectid is
used in order to get a stable ID for the object. This has all of the
usual costs of obtaining such an ID, though they aren't all that high
and this should be fairly unlikley on a hot path.