Introduce CompUnit::PrecompilationUnit
CompUnit::PrecompilationUnit represents precompiled bytecode and dependency
information as stored in a CompUnit::PrecompilationStore. It's a role that is
done by store specific classes. This interface allows for us to move much
knowledge about how a store is layed out back into the store's implementation
class.

This transition is not yet complete, as the repo-id is still stored in the
combined precomp file and the PrecompilationUnit cannot handle this extra bit
of information. Will be fixed by moving the repo-id into a separate file.