Make Junction.ACCEPTS 4x, Junction Bool 20x faster
The idea of using a scopeless STATEMENT_LIST is fine, but not if you're
starting a new scope for each iteration of the map.  On top of that, a
map needs to handle Slip's, and that only to be able to potentially
return from that scope.  This patch reforms the maps to a postfix
while with a conditional return in it.  So no new scope for each
iteration.  On top of that, if not returning on the way, we just set
the final value, rather than returning that: so no explicit return
in that case either.