Split off repo-id from precomp files
The repo-id is the identity of $*REPO (representing its contents and the whole
repo chain) at the time when a precomp file's dependencies were last resolved.
As long as the repo-id is the same, we can just load the dependencies by their
ids and do not have to resolve them from their DependencySpecifications.

When the repo-id is different, we need to re-resolve and check if the
dependencies still resolve to the same precomp files. If they do, we can still
use them and now store the updated repo-id in a precomp store. In other words:
the repo-id is volatile information compared to the byte code and the
dependency information.

Also, we probably don't need to protect this repo-id file with a lock, since
re-resolving dependencies in multiple processes is just a little wasteful but
harmless.