Speedup module loading by storing the result of dependency re-resolution
When $*REPO's identity (it's contents and all following repos) changed since we
wrote a precomp file, e.g. because we installed a dist, we have to re-resolve
all dependencies again, to find out if we need to re-compile.

In case we can continue to use the precomp file, we now store the updated
repo id in $*REPO's precomp store, so we only have to re-resolve when the repo
changes again.

panda's first run after installation takes 2.3 seconds while all following runs
now only take 1.8 seconds as we can just load the dependencies without
resolving them first.