Fix "Cannot call method 'prefix' on a null object" during JVM build
tools/build/install-core-dist.pl works with a repo chain consisting only of the
'perl' repository and missing 'vendor' and 'site'. This caused VMNull values in
the $custom-lib nqp hash over which name-for-repository stumbled later on.

Fix by removing those keys from the hash (if we would just leave them as is,
they'd contain the path strings).
With this the JVM build finishes successfully.