Store relative paths to dependency sources in precomp files if possible
We need the path of a dependency's source file to check the .modified time.
We also want to be able to move repositories without invalidating all
precomp files. So we use the same trick as for the source-names that we
use for backtraces and use paths relative to well known repositories.