Simplify $*USER/$*GROUP initialization
The whole IdFetch class is no longer necessary with the new way of
registering PROCESS:: dynamic variables.  Also make creation of IdName
objects faster with dedicated .new (although one could argue that for
the two objects that will ever exist for this, this isn't needed).

Unless of course, we're going to make the IdName class public as a
more generic "dualvar" class.