Fix debuggers hanging when the process needs to precompile
We try to use the same perl6 executable for precompilation as we use for
running the main program. That's not a good strategy if the main program
is running in a debugger as we'd start up a debugger for the precompilation,
too and those conflict on tty or tcp ports.