Fix trying to load an outdated precomp file
When a dependency of a dependency is outdated, we failed to recompile a
module, since we did detect it being outdated, tried to recompile but
the .precompile method thought someone else already got to it since it
only checked the modification time of the top level module.

Fix by correctly communicating the failure state up far enough to know
to pass :force.

Many thanks to llfourn++ for providing an excellent test case!
Fixes RT #128156