Use say for listing tarball contents
This one-liner to build our release tarballs was recently changed from using
Perl 5 to Perl 6; however, since Perl 6 autochomps unlike Perl 5,
the newlines are not preserved when we print(), causing the list of
filenames to run together.  We can use say() to emit newline after each
file to fix this.