Make iterating over a list about 10% faster again
Basically by not checking for the index being in range.  Since Lists
are immutable and cannot have holes (unlike arrays), just checking for
null is an equally good way to know you're done.

This of course assumes that the VM *does* check for index being in
range, but in a much faster way.