Eliminate &EXHAUST and the use of it.
We used to use it to replace the RETURN lexical, used for implementing
(lexotic) return. At one point, it would also give an improved error
for the case of returning from an already returned routine. However,
at this point it did not do so, and there are better ways to make that
happen. It did, however, carry a high cost, and frustrated inlining.
Its removal made one simple loop/methodcall/increment benchmark run in
63% of the CPU cycles it previously did, which is a nice win.