Extend handling of colonpair canon. in setting.
We'll soon be able to switch `return` to be a sub. However, doing so
fixes evaluating code that does a `return` while compiling the
setting. That in turn meant that we produced some better code at
compile-time for certain colonpairs, which in turn triggered an
unhandled case in canonicalization, which broke setting compilation.