Make object construction again 1.8x faster
It turns out that Mu.bless and Mu.BUILDALL were taking positional
parameters that were never actually given (apart from one arguably
faulty spectest).  This in turn caused flattening on an empty array
to be performed.  And then get passed to BUILDALL, which in turn
didn't do anything with it.  According to spec, bless should take
autovivifying type objects, but this has not been implemented.

Until that happens, it would seem we can use the performance boost
of not allowing positionals towards Mu.bless and Mu.BUILDALL, instead
of silently eating them and not doing anything with them.